C - Hello, World - 6. Size is not grandeur

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long int slli;
 	float sf;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(sc));
-	printf("Size of an int: %lu byte(s)\n", sizeof(si));
-	printf("size of a long int: %lu byte(s)\n", sizeof(sli));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(slli));
-	printf("Size of a float: %d byte(s)\n", sizeof(sf));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(sc));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(si));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(sli));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(slli));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(sf));
 	return (0);
 }
